Editor: moved several helper Control classes to AGS.Controls
