+++ Editor/AGS.Controls/Controls/SortableListView.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Windows.Forms;
 
-namespace AGS.Editor
+namespace AGS.Controls
 {
     /// <summary>
     /// SortableListView implements standard sorting of items using subitem
