+++ Editor/AGS.Controls/Controls/WinAPIHelper.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace AGS.Controls
+{
+    internal static class WinAPIHelper
+    {
+        public const uint WM_MOUSEACTIVATE = 0x21;
+        public const uint MA_ACTIVATE = 1;
+        public const uint MA_ACTIVATEANDEAT = 2;
+    }
+}
