+++ Editor/AGS.Editor/AGSEditor.csproj
@@ -157,7 +157,6 @@
     <Compile Include="Entities\CharacterRoomChangedEventArgs.cs" />
     <Compile Include="Entities\CompilationStepArgs.cs" />
     <Compile Include="Entities\CompileWarningEx.cs" />
-    <Compile Include="Entities\TreeItemDragEventArgs.cs" />
     <Compile Include="Entities\ExistingScriptHeaderToAdd.cs" />
     <Compile Include="Entities\ExistingFileToAdd.cs" />
     <Compile Include="Config\IObjectConfig.cs" />
@@ -168,8 +167,6 @@
     <Compile Include="Entities\OutputCreationStepArgs.cs" />
     <Compile Include="Entities\ScriptModuleDef.cs" />
     <Compile Include="Entities\GenericMessagesArgs.cs" />
-    <Compile Include="Entities\TreeItemTryDragEventArgs.cs" />
-    <Compile Include="Enums\TargetDropZone.cs" />
     <Compile Include="GUI\AssignToView.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -213,10 +210,6 @@
     <Compile Include="GUI\ImportTTFDialog.Designer.cs">
       <DependentUpon>ImportTTFDialog.cs</DependentUpon>
     </Compile>
-    <Compile Include="GUI\LineInBetween.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Include="GUI\ListViewExtensions.cs" />
     <Compile Include="GUI\LogPanel.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -248,9 +241,6 @@
     <Compile Include="GUI\ReplaceFolderDialog.Designer.cs">
       <DependentUpon>ReplaceFolderDialog.cs</DependentUpon>
     </Compile>
-    <Compile Include="GUI\SortableListView.cs">
-      <SubType>Component</SubType>
-    </Compile>
     <Compile Include="GUI\SplashScreen.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -280,10 +270,6 @@
     </Compile>
     <Compile Include="GUI\SpriteChooser.Designer.cs">
       <DependentUpon>SpriteChooser.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GUI\TreeViewExtensions.cs" />
-    <Compile Include="GUI\TreeViewWithDragDrop.cs">
-      <SubType>Component</SubType>
     </Compile>
     <Compile Include="GUI\ViewChooser.cs">
       <SubType>Form</SubType>
@@ -1038,9 +1024,6 @@
     <Compile Include="GUI\GotoLineDialog.Designer.cs">
       <DependentUpon>GotoLineDialog.cs</DependentUpon>
     </Compile>
-    <Compile Include="GUI\MenuStripExtended.cs">
-      <SubType>Component</SubType>
-    </Compile>
     <Compile Include="GUI\ProjectPanel.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -1064,9 +1047,6 @@
     </Compile>
     <Compile Include="GUI\SplashPage.Designer.cs">
       <DependentUpon>SplashPage.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GUI\ToolStripExtended.cs">
-      <SubType>Component</SubType>
     </Compile>
     <Compile Include="GUI\WindowsLayoutManager.cs" />
     <Compile Include="GUI\WindowsMenuManager.cs" />
