+++ Editor/AGS.Editor/GUI/ProjectTreeItem.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
 using System.Windows.Forms;
+using AGS.Controls;
 
 namespace AGS.Editor
 {
