+++ Editor/AGS.Editor/Panes/SpriteSelector.cs
@@ -10,6 +10,7 @@ using System.Drawing.Imaging;
 using System.IO;
 using System.Linq;
 using System.Windows.Forms;
+using AGS.Controls;
 
 namespace AGS.Editor
 {
